Fixed Code Review
- App과 index분리
- Button prors 변경
