+++ src/components/Buttons.jsx
@@ -6,7 +6,7 @@ function Buttons({ onClick }) {
   return (
     <p>
       {[1, 2, 3, 4, 5].map((number) => (
-        <Button key={number} number={number} onClick={() => onClick(number)}>
+        <Button key={number} onClick={() => onClick(number)}>
           {number}
         </Button>
       ))}
